+++ application/src/main/java/com/sap/cloud/s4hana/examples/HelloWorldServlet.java
@@ -1,0 +1,26 @@
+package com.sap.cloud.s4hana.examples;
+
+import org.slf4j.Logger;
+
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+import com.sap.cloud.sdk.cloudplatform.logging.CloudLoggerFactory;
+
+@WebServlet("/hello")
+public class HelloWorldServlet extends HttpServlet
+{
+    private static final long serialVersionUID = 1L;
+    private static final Logger logger = CloudLoggerFactory.getLogger(HelloWorldServlet.class);
+
+    @Override
+    protected void doGet( final HttpServletRequest request, final HttpServletResponse response )
+            throws IOException
+    {
+        logger.info("I am running!");
+        response.getWriter().write("Hello World!");
+    }
+}
